[PoC] Comp playground without MDX
