+++ ample-config.js
@@ -1,0 +1,5 @@
+import * as Button from "./src/components/button"
+
+export const components = {
+  button: Button
+}
