+++ package.json
@@ -21,6 +21,7 @@
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-react": "^7.18.0",
+    "faker": "^4.1.0",
     "fs-extra": "^8.1.0",
     "gatsby": "^2.20.14",
     "gatsby-image": "^2.2.34",
