+++ plugins/gatsby-ample-playground/src/templates/components/index.js
@@ -1,21 +1,27 @@
 import React from "react"
-import PropTypes from "prop-types"
-import { MDXRenderer } from "gatsby-plugin-mdx"
 import { Helmet } from "react-helmet"
+
+import Variations from "../../components/variations"
+
+import * as config from "../../../../../ample-config"
 
 import styles from "./styles.module.scss"
 
-import { getParentDir, getTitle } from "../../helpers"
+const ComponentsPlayground = () => {
+  const components = Object.entries(config.components).map((comp, idx) => {
+    const compName = comp[0]
+    const { component, fixtures } = comp[1]
+    const Component = component
 
-const ComponentsPlayground = ({ pageContext }) => {
-  const components = pageContext.components.map((comp, idx) => (
-    <div key={idx} className={styles.comp_section}>
-      <h2 id={getParentDir(comp.fileAbsolutePath)} className={styles.comp_heading}>
-        <a href={`#${getParentDir(comp.fileAbsolutePath)}`}>{getTitle(comp)}</a>
-      </h2>
-      <MDXRenderer>{comp.body}</MDXRenderer>
-    </div>
-  ))
+    return (
+      <div key={idx} className={styles.comp_section}>
+        <h2 id={compName} className={styles.comp_heading}>
+          <a href={`#${compName}`}>{component.name}</a>
+        </h2>
+        <Variations component={Component} data={fixtures} />
+      </div>
+    )
+  })
 
   return (
     <div className={styles.components_template}>
@@ -27,11 +33,7 @@ const ComponentsPlayground = ({ pageContext }) => {
   )
 }
 
-ComponentsPlayground.propTypes = {
-  pageContext: PropTypes.shape({
-    components: PropTypes.object.isRequired
-  }).isRequired
-}
+ComponentsPlayground.propTypes = {}
 
 ComponentsPlayground.defaultProps = {}
